Move lexer and parser
