+++ .gitignore
@@ -49,3 +49,6 @@ arxc
 # conda-forge
 conda/build/.ci_support
 conda/build/.scripts
+
+# IDE
+.vscode
