+++ arx/include/parser.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "lexer.h"
+
+namespace {
+class PrototypeAST;
+class ExprAST;
+}  // namespace
