+++ arx/src/arx.cpp
@@ -11,14 +11,13 @@
 // #include <arrow/status.h>
 // #include <arrow/table.h>
 
-#include "arx.h"
-
 #include <cctype>
 #include <cstdio>
 #include <iostream>
 #include <map>
 #include <string>
 #include <vector>
+
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/Analysis/BasicAliasAnalysis.h"
 #include "llvm/Analysis/Passes.h"
@@ -32,80 +31,12 @@
 #include "llvm/Support/TargetSelect.h"
 #include "llvm/Transforms/Scalar.h"
 
+#include "arx.h"
+#include "lexer.h"
+#include "parser.h"
+#include "utils.h"
+
 std::string ARX_VERSION = "1.1.1";  // semantic-release
-
-//===----------------------------------------------------------------------===//
-// Lexer
-//===----------------------------------------------------------------------===//
-
-// The lexer returns tokens [0-255] if it is an unknown character, otherwise
-// one of these for known things.
-enum Token {
-  tok_eof = -1,
-
-  // commands
-  tok_function = -2,
-  tok_extern = -3,
-  tok_return = -4,
-
-  // primary
-  tok_identifier = -10,
-  tok_number = -11,
-
-  // control
-  tok_if = -20,
-  tok_then = -21,
-  tok_else = -22,
-  tok_for = -23,
-  tok_in = -24,
-
-  // operators
-  tok_binary = -30,
-  tok_unary = -31,
-
-  // var definition
-  tok_var = -40,
-  tok_const = -41
-};
-
-std::string getTokName(int Tok) {
-  switch (Tok) {
-    case tok_eof:
-      return "eof";
-    case tok_function:
-      return "function";
-    case tok_return:
-      return "return";
-    case tok_extern:
-      return "extern";
-    case tok_identifier:
-      return "identifier";
-    case tok_number:
-      return "number";
-    case tok_if:
-      return "if";
-    case tok_then:
-      return "then";
-    case tok_else:
-      return "else";
-    case tok_for:
-      return "for";
-    case tok_in:
-      return "in";
-    case tok_binary:
-      return "binary";
-    case tok_unary:
-      return "unary";
-    case tok_var:
-      return "var";
-  }
-  return std::string(1, (char)Tok);
-}
-
-namespace {
-class PrototypeAST;
-class ExprAST;
-}  // namespace
 
 struct DebugInfo {
   llvm::DICompileUnit* TheCU;
@@ -116,89 +47,10 @@ struct DebugInfo {
   llvm::DIType* getDoubleTy();
 } KSDbgInfo;
 
-struct SourceLocation {
-  int Line;
-  int Col;
-};
-static SourceLocation CurLoc;
-static SourceLocation LexLoc = {1, 0};
-
-static int advance() {
-  int LastChar = getchar();
-
-  if (LastChar == '\n' || LastChar == '\r') {
-    LexLoc.Line++;
-    LexLoc.Col = 0;
-  } else
-    LexLoc.Col++;
-  return LastChar;
-}
-
-static std::string IdentifierStr;  // Filled in if tok_identifier
-static double NumVal;              // Filled in if tok_number
-
-/// gettok - Return the next token from standard input.
-static int gettok() {
-  static int LastChar = ' ';
-
-  // Skip any whitespace.
-  while (isspace(LastChar)) LastChar = advance();
-
-  CurLoc = LexLoc;
-
-  if (isalpha(LastChar)) {  // identifier: [a-zA-Z][a-zA-Z0-9]*
-    IdentifierStr = LastChar;
-    while (isalnum((LastChar = advance()))) IdentifierStr += LastChar;
-
-    if (IdentifierStr == "function") return tok_function;
-    if (IdentifierStr == "return") return tok_return;
-    if (IdentifierStr == "extern") return tok_extern;
-    if (IdentifierStr == "if") return tok_if;
-    if (IdentifierStr == "else") return tok_else;
-    if (IdentifierStr == "for") return tok_for;
-    if (IdentifierStr == "in") return tok_in;
-    if (IdentifierStr == "binary") return tok_binary;
-    if (IdentifierStr == "unary") return tok_unary;
-    if (IdentifierStr == "var") return tok_var;
-    return tok_identifier;
-  }
-
-  if (isdigit(LastChar) || LastChar == '.') {  // Number: [0-9.]+
-    std::string NumStr;
-    do {
-      NumStr += LastChar;
-      LastChar = advance();
-    } while (isdigit(LastChar) || LastChar == '.');
-
-    NumVal = strtod(NumStr.c_str(), nullptr);
-    return tok_number;
-  }
-
-  if (LastChar == '#') {
-    // Comment until end of line.
-    do LastChar = advance();
-    while (LastChar != EOF && LastChar != '\n' && LastChar != '\r');
-
-    if (LastChar != EOF) return gettok();
-  }
-
-  // Check for end of file.  Don't eat the EOF.
-  if (LastChar == EOF) return tok_eof;
-
-  // Otherwise, just return the character as its ascii value.
-  int ThisChar = LastChar;
-  LastChar = advance();
-  return ThisChar;
-}
-
 //===----------------------------------------------------------------------===//
 // Abstract Syntax Tree (aka Parse Tree)
 //===----------------------------------------------------------------------===//
 namespace {
-
-llvm::raw_ostream& indent(llvm::raw_ostream& O, int size) {
-  return O << std::string(size, ' ');
-}
 
 /// ExprAST - Base class for all expression nodes.
 class ExprAST {
